+++ stepic_plugins/quizzes/admin/__init__.py
@@ -7,9 +7,10 @@ import time
 
 import requests
 
+from stepic_plugins import settings
 from stepic_plugins.base import BaseQuiz
 from stepic_plugins.exceptions import FormatError, PluginError
-from stepic_plugins.quizzes.executable_base import jail_code_wrapper, settings
+from stepic_plugins.executable_base import jail_code_wrapper
 
 
 logger = logging.getLogger(__name__)
